Add thanks Sanket
This is just a trivial patch to recognize that the code here was
originally hacked together by Sanket, and so we can get an ack on github
to show he is okay with me stealing the code and ok with me leaving his
name in the author list.

Thanks Sanket!